Refactoring: Moved reporting request handlers from serviceProviderReferralsController.ts into it's own class called reportingController.ts, as well as moving it closer to reporting Presenter/View classes.
The serviceProviderReferralsController.ts and it's equivalent test class has too many method defined, this has the consequence of:

1. More chances of conflicts with other commits when making changes
2. Harder to read and locate code.
3. Overloaded constructor
